+++ Integrations/Guinevere.Vulkan.SilkNET/GuiWindow.cs
@@ -37,6 +37,7 @@ public unsafe class GuiWindow : IInputHandler, IWindowHandler, IDisplayCapabilit
     readonly StringBuilder _typedCharacters = new();
     readonly Font _fontText;
     readonly Font _fontIcon;
+    readonly Font _fontWidgetIcon;
     readonly Glfw _glfw = Glfw.GetApi();
 
     /// <summary>
@@ -60,6 +61,8 @@ public unsafe class GuiWindow : IInputHandler, IWindowHandler, IDisplayCapabilit
         _fontText = Font.FromStream(fontStream);
         fontStream = GetStreamResource("Guinevere.icons.ttf");
         _fontIcon = Font.FromStream(fontStream);
+        fontStream = GetStreamResource("Guinevere.widget-icons.ttf");
+        _fontWidgetIcon = Font.FromStream(fontStream);
         _renderer = new CanvasRenderer(logger);
         _gui.Platform.Register<ICanvasRenderer>(_renderer);
 
@@ -182,7 +185,7 @@ public unsafe class GuiWindow : IInputHandler, IWindowHandler, IDisplayCapabilit
                 try
                 {
                     _gui.SetStage(Pass.Pass1Build);
-                    _gui.BeginFrame(canvas, _fontText, _fontIcon);
+                    _gui.BeginFrame(canvas, _fontText, _fontIcon, _fontWidgetIcon);
                     _draw();
 
                     // Process the whole layout after the build pass
@@ -293,6 +296,7 @@ public unsafe class GuiWindow : IInputHandler, IWindowHandler, IDisplayCapabilit
         _window.Dispose();
         _fontText.Dispose();
         _fontIcon.Dispose();
+        _fontWidgetIcon.Dispose();
         GC.SuppressFinalize(this);
     }